pyproject.toml
[tool.poetry]
name = "ultra-ssdbox-manager"
version = "0.1.0"
description = "GUI Manager for R-SODIUM Ultra SSDBox"
authors = ["Barryblueice <barryblueice@outlook.com>"]
license = "MPL-2.0"
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyqt6 = "6.4.2"
pyqt6-tools = "^6.4.2.3.3"
windows-toasts = "^1.3.1"
pyqt-toast-notification = "^1.3.3"
pywin32 = "^310"
usb = "^0.0.83.dev0"
pyusb = "^1.3.1"


[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
